CW-44 Merge feedback_test.go into mailing_test.go
